refactor: extract engine-cli-boundary gate into scripts/engine_cli_boundary.py
check_ai_readiness.py was already past the 2000-line hard cap and only
green via LARGE_FILE_ALLOWLIST -- mirrors the adr_status_sync.py
extraction already established for the same reason. Purely mechanical:
the check's behavior, allowlist, and docstrings are unchanged, only the
module they live in.

Also, in response to review:
- docs: fix the eleven-vs-ten user-facing-operations inconsistency (the
  probe harness, item 8, was both counted in "eleven" and explicitly
  excluded from "the eleven user-facing operations" a paragraph later)
- docs: correct the Phase 0 item 1 module predicate from service_* to
  service*.py (including bare service.py), matching _is_engine_module
- AGENTS.md, check docstring/error text: add artifact_*.py to the
  engine-cli-boundary scope description (the predicate already covered
  it; only the prose didn't)
- docs: split the artifact-resolution equivalence test's field list into
  a full-resolution tier and a narrower symbols_only=True tier, since a
  symbols_only call never computes header/L2-derived fields
- docs: add a Phase 2 step routing appcompat's two direct
  compare_snapshots() calls and deps compare's direct checker.compare()
  call onto EffectiveEvaluationConfig, since no other listed step
  reaches them and the comparison-equivalence acceptance test needs it
