+++ scripts/CLAUDE.md
@@ -13,6 +13,7 @@ Each must run with Python 3.10+ and the package installed in dev mode
 | `gen_repo_facts.py` | Single source of truth for volatile repository facts (CLAUDE.md "M1-4") — project/latest-release version, example-catalog size, fast-lane test count, canonical Python. Writes `repo_facts.json`; `--check` fails on drift. `latest_release` is a maintained field (bump it when cutting a release), not auto-derived from git tags (unreliable on a shallow CI checkout). | `verify.py --only repo-facts` (CI `ai-readiness` job); manual. |
 | `check_ai_readiness.py` | AI-readiness gate (file size, CLAUDE.md coverage, test ratio, ChangeKind invariants, mypy baseline drift, import cycles, engine/CLI dependency-direction boundary, test-assertion density, ADR index/nav + status sync). | CI (`ai-readiness`) and `pre-commit`, both via `verify.py --only ai-readiness`. Exits 1 on errors. |
 | `adr_status_sync.py` | The `adr-status-sync` gate, split out of `check_ai_readiness.py` (already past the 2000-line hard cap): an ADR's own `**Status:**` line and its row in `docs/contribute/adr/index.md` may not *contradict* each other, an optional `**Verified:** <ref>@<sha> on <date>` receipt must parse, and commits after that sha touching a first-party file the Status names raise a WARN — the only check here that catches *document-vs-code* drift. Also owns the ADR Status parsing helpers (`_ADR_FILE_RE`, `_adr_status_text`) that `check_ai_readiness.py`'s own `adr-index-nav-sync` reuses. Pure-stdlib, importable. | imported (by `check_ai_readiness.py`); gated via `tests/test_ai_readiness.py`. |
+| `engine_cli_boundary.py` | The `engine-cli-boundary` gate, split out of `check_ai_readiness.py` the same way `adr_status_sync.py` was (same 2000-line-cap reason): no engine-layer module (`scan_engine.py`, `service*.py`, `artifact_*.py`, `buildsource/**/*.py`) may import `click` or a `cli`/`cli_*` sibling — including a *nested* CLI adapter like `abicheck/compat/cli.py` — allowlist-and-shrink (`ENGINE_CLI_BOUNDARY_ALLOWLIST`), keyed by a stable `<rel>::<desc>::<occurrence>` identity rather than line number. A trailing imported alias is only flagged once verified against the real file tree as an actual submodule (`_alias_is_real_submodule`), so an ordinary symbol that merely starts with `cli_` is never a false positive. Pure-stdlib, importable. | imported (by `check_ai_readiness.py`); gated via `tests/test_engine_cli_boundary.py`. |
 | `findings_report.py` | The shared `Findings` error/warning collector both gate scripts report through — grouping by check name, the `ERROR:`/`WARN:` rendering, and the `<label>: N error(s), M warning(s)` summary line. `check_ai_readiness.py` and `check_docs_contract.py` each subclass it to pin their own label; previously each carried a byte-identical copy of the class. Pure-stdlib (importable before `pip install -e .`, same constraint as `check_ai_readiness.py` itself). | imported (by `check_ai_readiness.py`, `check_docs_contract.py`). |
 | `check_fp_rate.py` | False-positive/false-negative gate for public-surface scoping (ADR-024 §7). Labelled `(old, new)` corpus; baselines FP=0/FN=0. Cases are tagged by scoping *axis* (`CASE_CATEGORY`): `--json` carries a `by_category` breakdown and `--markdown` renders a per-axis table for CI step-summary / release-over-release trend reading. | CI (`ai-readiness`). Mirrored in `tests/test_fp_rate_gate.py`. |
 | `check_tier_accuracy.py` | Per-evidence-tier accuracy gate — *what each level buys*. One labelled logical change per case, projected down to what each tier observes (L0 symbols → L1 debug → L2 headers → L3 build) and run through `compare`; verdicts collapse to a 3-band ordinal (non-breaking/risk/breaking). Measures per-tier **over-call (FP)** vs **under-call (FN)** and which transition removes each. Gates: top-tier correctness + under-call monotonicity (more evidence never hides a break — ADR-028 D3). `--markdown`/`--json` emit the per-tier matrix. | CI (`ai-readiness`, + step-summary artifact). Mirrored in `tests/test_tier_accuracy_gate.py`. |
